refactor: moved some files for better cohesiveness in file groups
