+++ src/components/modals/DrawerModal.tsx
@@ -17,8 +17,8 @@ import { useEmojiStore } from "../../store/globalStore";
 import { EmojiGroup } from "../../types/EmojiData";
 import { IProject } from "../../types/ProjectTypes";
 import { DrawerModalFormValues } from "../../types/TopLeftBarTypes";
-import useProjectIcon from "../../utils/useProjectIcon";
-import AsyncEmojiPicker from "../AsyncEmojiPicker";
+import useProjectIcon from "../../hooks/useProjectIcon";
+import AsyncEmojiPicker from "../ui/AsyncEmojiPicker";
 import { ContextModalProps } from "@mantine/modals";
 import { StatusIcon } from "../icons/StatusIcon";
 import { isSuccessStatus, determineTitle } from "../../utils/successHelpers";
