+++ src/components/ui/ConditionalHoverCard.tsx
