+++ src/hooks/useEmojiData.tsx
