+++ src/hooks/useIsTruncated.tsx
